+++ src/main/java/com/github/kryvii/lushgrass/client/LushGrassClient.java
@@ -1,24 +1,15 @@
 package com.github.kryvii.lushgrass.client;
 
-import com.github.kryvii.lushgrass.LushGrass;
 import com.github.kryvii.lushgrass.client.event.ClientConfigEvents;
 import com.github.kryvii.lushgrass.client.event.ClientModelEvents;
-import com.github.kryvii.lushgrass.config.ClientConfig;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.IEventBus;
-import net.neoforged.fml.ModContainer;
-import net.neoforged.fml.common.Mod;
-import net.neoforged.fml.config.ModConfig;
-import net.neoforged.neoforge.client.gui.ConfigurationScreen;
-import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import net.minecraftforge.eventbus.api.IEventBus;
 
-@Mod(value = LushGrass.MOD_ID, dist = Dist.CLIENT)
 public final class LushGrassClient {
-    public LushGrassClient(IEventBus modEventBus, ModContainer container) {
-        container.registerConfig(ModConfig.Type.CLIENT, ClientConfig.CLIENT_SPEC);
-        container.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
-
+    public static void register(IEventBus modEventBus) {
         ClientModelEvents.register(modEventBus);
         ClientConfigEvents.register(modEventBus);
+    }
+
+    private LushGrassClient() {
     }
 }