+++ src/main/java/com/github/kryvii/lushgrass/client/event/ClientConfigEvents.java
@@ -2,9 +2,9 @@ package com.github.kryvii.lushgrass.client.event;
 
 import com.github.kryvii.lushgrass.LushGrass;
 import net.minecraft.client.Minecraft;
-import net.neoforged.bus.api.IEventBus;
-import net.neoforged.fml.config.ModConfig;
-import net.neoforged.fml.event.config.ModConfigEvent;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.config.ModConfig;
+import net.minecraftforge.fml.event.config.ModConfigEvent;
 
 public final class ClientConfigEvents {
     public static void register(IEventBus modEventBus) {